Create @lexical/html package
