+++ packages/lexical-clipboard/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@lexical/utils": "0.2.9",
     "@lexical/selection": "0.2.9",
-    "@lexical/headless": "0.2.9"
+    "@lexical/html": "0.2.9"
   },
   "repository": {
     "type": "git",
