+++ packages/lexical-clipboard/src/__tests__/unit/clipboard.test.ts
@@ -6,7 +6,7 @@
  *
  */
 
-import {$generateHtmlFromNodes, $generateNodesFromDOM} from '@lexical/headless';
+import {$generateHtmlFromNodes, $generateNodesFromDOM} from '@lexical/html';
 import {$createLinkNode, LinkNode} from '@lexical/link';
 import {
   $createListItemNode,
