+++ packages/lexical-html/flow/LexicalHtml.js.flow
@@ -1,0 +1,28 @@
+/**
+ * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
+ */
+
+import type {
+  LexicalEditor,
+  LexicalNode,
+  EditorState,
+  EditorThemeClasses,
+  RangeSelection,
+  NodeSelection,
+  GridSelection,
+} from 'lexical';
+
+declare export function $generateHtmlFromNodes(
+  editor: LexicalEditor,
+  selection?: RangeSelection | NodeSelection | GridSelection | null,
+): string;
+
+declare export function $generateNodesFromDOM(
+  editor: LexicalEditor,
+  dom: Document,
+): Array<LexicalNode>;
