+++ packages/lexical-html/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@lexical/html",
+  "description": "This package contains HTML helpers and functionality for Lexical.",
+  "keywords": [
+    "lexical",
+    "editor",
+    "rich-text",
+    "html"
+  ],
+  "license": "MIT",
+  "version": "0.2.9",
+  "main": "LexicalHtml.js",
+  "peerDependencies": {
+    "lexical": "0.2.9"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/facebook/lexical",
+    "directory": "packages/lexical-html"
+  }
+}
