+++ tsconfig.json
@@ -24,6 +24,7 @@
       "@lexical/offset": ["./packages/lexical-offset/LexicalOffset.d.ts"],
       "@lexical/history": ["./packages/lexical-history/LexicalHistory.d.ts"],
       "@lexical/headless": ["./packages/lexical-headless/LexicalHeadless.d.ts"],
+      "@lexical/html": ["./packages/lexical-html/LexicalHtml.d.ts"],
       "@lexical/utils": ["./packages/lexical-utils/LexicalUtils.d.ts"],
       "@lexical/code": ["./packages/lexical-code/LexicalCode.d.ts"],
       "@lexical/plain-text": [
